requirements.txt
pydle
isodate
pyyaml
httplib2
oauth2client
google-api-python-client
